DjangoGram/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cloudinary==1.32.0
coverage==7.2.2
Django==4.1.7
django-environ==0.10.0
Pillow==9.4.0
psycopg2-binary==2.9.6
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
